workspace/2016 TCO Algorithm/EllysTimeMachine.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Numerics;
using Debug = System.Diagnostics.Debug;
using StringBuilder = System.Text.StringBuilder;
public class EllysTimeMachine
{
    public string getTime(string time)
    {
        var h = (time[0] - '0') * 10 + time[1] - '0';
        var t = (time[3] - '0') * 10 + time[4] - '0';
        Debug.WriteLine(h);
        Debug.WriteLine(t);
        if (t == 0) t = 60;
        if (h == 12) h = 0;
        var s = string.Format("{0:00}:{1:00}", t / 5, h * 5);
        return s;
    }

    static public T[] Enumerate<T>(int n, Func<int, T> f) { var a = new T[n]; for (int i = 0; i < n; ++i) a[i] = f(i); return a; }
    static public void Swap<T>(ref T a, ref T b) { var tmp = a; a = b; b = tmp; }
}
static public class EnumerableEX
{
    static public string AsString(this IEnumerable<char> e) { return new string(e.ToArray()); }
    static public string AsJoinedString<T>(this IEnumerable<T> e, string s = " ") { return string.Join(s, e); }
}
// CUT begin
public class Tester : AbstractTester
{
    static public T[] Enumerate<T>(int n, Func<int, T> f) { var a = new T[n]; for (int i = 0; i < n; ++i) a[i] = f(i); return a; }
    public Random rand = new Random(0);
    public Tester()
    {
    }
}
// CUT end
